slang pre2.1.4 changes 52-53 52. src/slang.h: Added SLFUTURE_CONST to arg_type field of     SLcmd_Cmd_Type object. 53. src/slarith.c: The return value of fmod was being truncated to     single precision.
git-svn-id: http://opensvn.csie.org/jedsoft/slang/trunk@248 93520148-bd12-0410-9693-835737c2d474

+++ src/slarith.c
@@ -362,7 +362,7 @@ SLarith_get_to_double_fun (SLtype type, unsigned int *sizeof_type)
 #define GENERIC_TYPE double
 #define POW_FUNCTION(a,b) pow((double)(a),(double)(b))
 #define POW_RESULT_TYPE double
-#define MOD_FUNCTION(a,b) (float)fmod((a),(b))
+#define MOD_FUNCTION(a,b) fmod((a),(b))
 #define TRAP_DIV_ZERO	0
 #define GENERIC_UNARY_FUNCTION double_unary_op
 #define GENERIC_ARITH_UNARY_FUNCTION double_arith_unary_op
